model.touch when using short ts
Making touch method update the updated_at field, even when using alias

+++ spec/app/models/label.rb
@@ -1,6 +1,6 @@
 class Label
   include Mongoid::Document
-  include Mongoid::Timestamps::Updated
+  include Mongoid::Timestamps::Updated::Short
 
   field :name, type: String
   field :after_create_called, type: Boolean, default: false
